dataset_pipeline/requirements.txt
zstandard
pmaw
praw
pandas
openai
weaviate-client